requirements.txt
django==1.10.3
dj-pagination
pymysql
jinja2
django-celery==3.2.2
celery==3.1.26.post2
djangorestframework
redis==2.10.6
ansible==2.7.9
requests==2.21.0